test: refactor to eliminate flaky test
This retains the key elements of test-child-process-fork-getconnections
(forks a child process, sends a bunch of sockets, uses getConnections()
to enumerate them) but contains some code to work around an apparent
intermittent bug that occurs on OS X where a socket seems to close
itself unexpectedly.

https://github.com/nodejs/node/issues/2610 was opened for the bug that
was causing the problem in the first place.

PR-URL: https://github.com/nodejs/node/pull/2609
Fixes: https://github.com/nodejs/node/issues/1100

Fixes: 1100
PR-URL: https://github.com/nodejs/node/pull/2609
Reviewed-By: brendanashworth - Brendan Ashworth <brendan.ashworth@me.com>
Reviewed-By: jbergstroem - Johan Bergström <bugs@bergstroem.nu>